+++ lib/Brani/src/Batch/BraniSubscribeWebhooks.ts
@@ -20,7 +20,11 @@ export default class BraniSubscribeWebhooks extends ABatchNode {
         const app = this.getApplication<BraniApplication>();
         const appInstall = await this.getApplicationInstallFromProcess(dto);
 
-        const registered = await repo.findMany({ users: [appInstall.getUser()], apps: [appInstall.getName()] });
+        const registered = await repo.findMany({
+            users: [appInstall.getUser()],
+            apps: [appInstall.getName()],
+            sdks: [appInstall.getSdk()],
+        });
         const all = app.getWebhookSubscriptions().map((sub) => ({
             event: sub.getName(),
             token: this.getRandomToken(),
@@ -42,6 +46,7 @@ export default class BraniSubscribeWebhooks extends ABatchNode {
             const wb = new Webhook()
                 .setWebhookId('n/a')
                 .setUser(appInstall.getUser())
+                .setSdk(appInstall.getSdk())
                 .setNode(unregistered.node)
                 .setToken(unregistered.token)
                 .setApplication(app.getName())