+++ lib/Shopify/src/Batch/ShopifyRegisterWebhook.ts
@@ -54,6 +54,7 @@ export default class ShopifyRegisterWebhook extends ABatchNode {
                 await repo.insert(new Webhook()
                     .setWebhookId(respBody.webhook.id.toString())
                     .setUser(appInstall.getUser())
+                    .setSdk(appInstall.getSdk())
                     .setNode(webhooks[webhookIndex].getNode())
                     .setToken(token)
                     .setApplication(app.getName())