+++ lib/Wflow/src/Batch/WflowUnsubscribeWebhookBatch.ts
@@ -4,6 +4,7 @@ import ABatchNode from '@orchesty/nodejs-sdk/dist/lib/Batch/ABatchNode';
 import OnRepeatException from '@orchesty/nodejs-sdk/dist/lib/Exception/OnRepeatException';
 import { HttpMethods } from '@orchesty/nodejs-sdk/dist/lib/Transport/HttpMethods';
 import BatchProcessDto from '@orchesty/nodejs-sdk/dist/lib/Utils/BatchProcessDto';
+import { StatusCodes } from 'http-status-codes';
 import WflowApplication, { NAME as WFLOW_APP_NAME } from '../WflowApplication';
 
 export const NAME = `${WFLOW_APP_NAME}-unsubscribe-webhooks-batch`;
@@ -22,6 +23,7 @@ export default class WflowUnsubscribeWebhookBatch extends ABatchNode {
         const webhook = await repository.findOne({
             users: [appInstall.getUser()],
             apps: [appInstall.getName()],
+            sdks: [appInstall.getSdk()],
         });
 
         if (!webhook) {
@@ -45,7 +47,7 @@ export default class WflowUnsubscribeWebhookBatch extends ABatchNode {
 
         const response = await this.getSender().send(request);
 
-        if (response.getResponseCode() as number !== 200) {
+        if (response.getResponseCode() !== StatusCodes.OK) {
             await repository.update(webhook.setUnsubscribeFailed(true));
             throw new OnRepeatException(300, 12, response.getBody());
         }